Rename "R5DEV" --> "DEVELOPER" see description ( !!big change!! )
The "DEV" global define was never set in this retail engine, even when the engine was launched with "-dev", or when "developer" ConVar was set, and the ScriptVM was reset. It caused confusion because the DEV define DOES work in the RSON files. In the early days of this project (before it was released), we renamed all DEV macro's to R5DEV and set this global const to true in init.nut. This was a dirty solution, because it forces all development scripts, and even caused compile errors when the game was not launched with "-dev", because the RSON files WON'T compile anything under DEV blocks, but the scripts do expect them as R5DEV was always true. The SDK registers a constant called "DEVELOPER" with the value of the developer ConVar. The reason it had to be renamed from "R5DEV" to "DEVELOPER" was to avoid confusion with left over Jira tickets, the reason we couldn't use "DEV" was because it's already defined, but as NULL. Using all the effort to perform inline asm hooking to swap the value with that of the "developer" ConVar was not worth it.