Survival Commentary Major Update Announcements are now announced in the same way as in the original game. (Specify eDialogueFlags.SURVIVAL_HOST_ALL_SPEAKERS for the dialogFlags in PlayDialogueForPlayer)
This allows the announcements to be heard in three dimensions, as they are made through the environmental speakers (like the vertical display that regularly shows champions and kill leaders).

Added legend responses to announcements. (WIP, I will update this in the future)

Changed the announcements to be played to be taken from the eSurvivalCommentaryBucket. This eliminates the need to directly describe the sound to be played.

Flowstate:
The announcement for multiple kills by the kill leader is temporarily disabled because it does not exist in the dialogue. I will create another way to play it.