Implement Flowstate_CheckPlayerIsAFK
If gamemode is custom_tdm or custom_ctf, a thread is started when player is fully connected, this thread checks player's velocity every second. If player stand still for 180 seconds (3 mins), player will be kicked from the server, a message alerting the player will display 10 seconds before and if player moves the timer is restarted. This feature was requested by the Chinese community and is also useful for every flowstate server.

It checks if player is not spectating, if player is alive and if eGameState is playing as conditions to start the afk timer.