temp solution for lifeline drone being solid + hovering
in retail it does not hover. You can test it by placing it while falling from a building, in r5r it uses physics mover and it will drop slowly, this will also fix that, however it won't bob locally anymore, in retail it does bob locally. todo find a solution for the bob. todo find a solution for not sticking to train anymore (but i do prefer this for now)