Move OnPlayerDamaged liveapi event callback to OnPostDamaged instead
There was a report of the OnPlayerDamaged event callback sending up incorrect results as there is additional logic that runs after this callback that alters the final damage that gets applied to the player. Removed the old call and installed the callback directly on OnPostDamaged to see if this will yield correct results. The change has been tested and appears to work good so far, needs more testing.