fix ammo spawning
save edicts by a lot
based on https://www.ea.com/games/apex-legends/news/space-hunt-event

"Ammo that spawns with a weapon will now always be a single brick, but with twice the ammo"
Dev Note: Previously, there was a 70% chance to spawn two separate ammo bricks and a 30% chance to only spawn one when a weapon spawned with its corresponding ammo. We’ve changed this to clean up the number of items on the ground and slightly increase the ammo economy.