READ DESCRIPTION ( no really, read it. )
NOTICE: Updated defaults
NOTICE: If using tracker, updated dll

- Added localization strings for all languages
- player struct vars for mute data
- cl stats remote funcs
- stat util functions
- Gamemode must explicitely use Gamemode1v1_SetRestEnabled if using 1v1/3v3rest system
- Fixed rest for scenarios.
- Added ui confirm for rest during match
- Added penalty to score for resting during match
- Fixed issue with tracker waiting for flag that which the setter was never called
- Wrapper SpawnSystem_SetPanelLocation
- Updated antiafk logic
- Client stats framework: Allows you to get the stats of other players on an as need basis without using netbools etc
- New stats function helpers
- New backend "read_only" class stats, updated only in backend cms helper portal
- Commands/args registration to easily streamline commands/args and their accepted aliases per gamemode/playlist

- New mute system - with additionalcc commands
[
cc mute playername/id -r "this is a reason" 1 day 5 hours 10 seconds

Any combination of time strings can be added, and the order of parameters does not matter,

cc mute playername/id  10 seconds -reason "this is a reason"

is also valid

If a player is muted with a timestring that evaluates to a time that is within the current match, a thread is dispatched to unmute them at that time.

This also happens when a player joins.

It is possible to mute/unmute players by uid even if they are offline

- Opt in global mute: players muted by helpers globally will auto mute in your server

- opt in spam mute: configure opt in spam muting thresholds and vars in playlists
]

- fixed some stats damage defs
- minor fixes and refactors